backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
gunicorn
whitenoise
psycopg2-binary
